requirements.txt
torch>=2.0.0
torchvision
diffusers>=0.25.0
transformers
accelerate
safetensors
fastapi
uvicorn[standard]
pillow
opencv-python
runpod


